Added some useful Stuff
In this small patch I just added a Config-Loader from an other project and migrated it for this one and did created some classes which need to be coded soon. On top of that I added a complex datatype called Vector (will be improved soon) which is a representation of a Vector in a 3D room.