Mayor Patch with sensor support
In this patch I tried to implement the accelaration-sensor, the gyroscope and the temperature-sensor which may not be needed but is included in the sensor which was ordered.
In order to add multi-sensor support I will maybe add a config for the registers soon. At the moment you still have to change everything just in the code.

Other patchnotes:
*Fixed some obvious bugs

IMPORTANT: Code is not tested as the sensor hasn't already arrived.